+++ jobbergate-cli/CHANGELOG.md
@@ -3,6 +3,9 @@
 This file keeps track of all notable changes to jobbergate-cli
 
 ## Unreleased
+
+
+## 5.3.0a5 -- 2024-08-30
 - Enhance error message when request returns 403 Forbidden [ASP-5433]
 
 ## 5.3.0a4 -- 2024-08-23